Add PostPublisher Jekyll Plugin
Added a jekyll plugin called PostPublisher. 99% of code comes from a blog post
over at [Jeffrey Sambells][jsambells] blog. The code as-is was crashing when
"date: unpublished" existed in the yaml header in unpublished posts, so some
tweaks were made. The plugin now yanks out the yaml header and injects a date
field (overwriting it if it exists) instead of assuming the date field exists
and gsubbing it.

[jsambells]: http://http://jeffreysambells.com/2013/02/01/jekyll-draft-publishing-plugin/