Add new_post script
Wrote a ruby script called `new_post`, which generates a new draft post and
creates a new file in _drafts. It can be invoked in one of two ways:

  - ./new_post Node Is Just The Coolest Thing Ever
  - ./new_post "Node Is Just The Coolest Thing Ever"

In either case, new_post will do the intelligent thing, preserving formatting
for your actual title while downcasing and concating the filename into the
format: "node-is-just-the-coolest-thing-ever.md".

It also properly generates default YAML frontmatter and inserts it into the
created file. All that's left is to write the blog post, and then drop the
finished draft into `_publish` when it needs to be timestamped and posted.